refactor: change Icons component to Icon
