+++ src/app/(auth)/sign-in/index.tsx
@@ -2,8 +2,8 @@ import { StyleSheet, Text, TouchableOpacity, View } from 'react-native'
 import { StatusBar } from 'expo-status-bar'
 import { useTranslation } from 'react-i18next'
 
+import { IconStar } from '@components/icon'
 import IconInformation from '@components/icon_information'
-import { IconStar } from '@components/icons'
 import { useTheme } from '@contexts'
 import { useAuth } from '@contexts'
 import changeLanguage from '@locales/changeLanguage'
