+++ src/app/(tabs)/_layout.tsx
@@ -1,6 +1,6 @@
-import Calendar from '@components/icons/calendar'
-import Home from '@components/icons/home'
-import Person from '@components/icons/person'
+import Calendar from '@components/icon/calendar'
+import Home from '@components/icon/home'
+import Person from '@components/icon/person'
 import NavBar from '@components/nav_bar'
 
 const _layout = (): JSX.Element => {
