+++ src/components/icon/star/index.tsx
