+++ src/components/icon/types.ts
